Extract the fixed-effects estimates from a fitted random effects model

A generic function to extract the fixed effects (i.e. the company-specific effects) estimates from a fitted random effects model.

Usage

# S3 method for class 'hierCredGLM'
fixef(object, ...)

# S3 method for class 'hierCredTweedie'
fixef(object, ...)

Arguments

object

an object of type hierCredGLM or hierCredTweedie

...

ignored.

Value

a named, numeric vector of fixed-effects estimates.

Examples

# \donttest{
fit = hierCredGLM(Y ~ area + (1 | VehicleType / VehicleBody), dataCar,
weights = w, p = 1.75, epsilon = 1e-6)
#> Error in hierCredGLM(Y ~ area + (1 | VehicleType/VehicleBody), dataCar,     weights = w, p = 1.75, epsilon = 1e-06): Did not find the variables in the formula in the dataframe
fixef(fit)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'object' in selecting a method for function 'fixef': object 'fit' not found
# }