Default Printing

print.default is the default method of the generic print function which prints its argument.

Usage

# S3 method for BalanceProperty
print(x, ...)

Arguments

x

the object to be printed.

...

further arguments to be passed to or from other methods. They are ignored in this function.

Details

The default for printing NAs is to print NA (without quotes) unless this is a character NA and quote = FALSE, when <NA> is printed.

The same number of decimal places is used throughout a vector. This means that digits specifies the minimum number of significant digits to be used, and that at least one entry will be encoded with that minimum number. However, if all the encoded elements then have trailing zeroes, the number of decimal places is reduced until at least one element has a non-zero final digit. Decimal points are only included if at least one decimal place is selected.

Attributes are printed respecting their class(es), using the values of digits to print.default, but using the default values (for the methods called) of the other arguments.

Option width controls the printing of vectors, matrices and arrays, and option deparse.cutoff controls the printing of language objects such as calls and formulae.

When the methods package is attached, print will call show for R objects with formal classes (‘S4’) if called with no optional arguments.

Large number of digits

Note that for large values of digits, currently for digits >= 16, the calculation of the number of significant digits will depend on the platform's internal (C library) implementation of sprintf() functionality.

Single-byte locales

If a non-printable character is encountered during output, it is represented as one of the ANSI escape sequences (\a, \b, \f, \n, \r, \t, \v, \\ and \0: see Quotes), or failing that as a 3-digit octal code: for example the UK currency pound sign in the C locale (if implemented correctly) is printed as \243. Which characters are non-printable depends on the locale. (Because some versions of Windows get this wrong, all bytes with the upper bit set are regarded as printable on Windows in a single-byte locale.)

Multi-byte locales

In all locales, the characters in the ASCII range (0x00 to 0x7f) are printed in the same way, as-is if printable, otherwise via ANSI escape sequences or 3-digit octal escapes as described for single-byte locales. Whether a character is printable depends on the current locale and the operating system (C library).

Multi-byte non-printing characters are printed as an escape sequence of the form \uxxxx (in hexadecimal). This is the Unicode code point of the character.

It is possible to have a character string in a character vector that is not valid in the current locale. If a byte is encountered that is not part of a valid character it is printed in hex in the form \xab and this is repeated until the start of a valid character.

See also

Examples

pi
#> [1] 3.141593
print(pi, digits = 16)
#> [1] 3.141592653589793
LETTERS[1:16]
#>  [1] "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "N" "O" "P"
print(LETTERS, quote = FALSE)
#>  [1] A B C D E F G H I J K L M N O P Q R S T U V W X Y Z

M <- cbind(I = 1, matrix(1:10000, ncol = 10,
                         dimnames = list(NULL, LETTERS[1:10])))
utils::head(M)        # makes more sense than
#>      I A    B    C    D    E    F    G    H    I    J
#> [1,] 1 1 1001 2001 3001 4001 5001 6001 7001 8001 9001
#> [2,] 1 2 1002 2002 3002 4002 5002 6002 7002 8002 9002
#> [3,] 1 3 1003 2003 3003 4003 5003 6003 7003 8003 9003
#> [4,] 1 4 1004 2004 3004 4004 5004 6004 7004 8004 9004
#> [5,] 1 5 1005 2005 3005 4005 5005 6005 7005 8005 9005
#> [6,] 1 6 1006 2006 3006 4006 5006 6006 7006 8006 9006
print(M, max = 1000)  # prints 90 rows and a message about omitting 910
#>         I    A    B    C    D    E    F    G    H    I     J
#>    [1,] 1    1 1001 2001 3001 4001 5001 6001 7001 8001  9001
#>    [2,] 1    2 1002 2002 3002 4002 5002 6002 7002 8002  9002
#>    [3,] 1    3 1003 2003 3003 4003 5003 6003 7003 8003  9003
#>    [4,] 1    4 1004 2004 3004 4004 5004 6004 7004 8004  9004
#>    [5,] 1    5 1005 2005 3005 4005 5005 6005 7005 8005  9005
#>    [6,] 1    6 1006 2006 3006 4006 5006 6006 7006 8006  9006
#>    [7,] 1    7 1007 2007 3007 4007 5007 6007 7007 8007  9007
#>    [8,] 1    8 1008 2008 3008 4008 5008 6008 7008 8008  9008
#>    [9,] 1    9 1009 2009 3009 4009 5009 6009 7009 8009  9009
#>   [10,] 1   10 1010 2010 3010 4010 5010 6010 7010 8010  9010
#>   [11,] 1   11 1011 2011 3011 4011 5011 6011 7011 8011  9011
#>   [12,] 1   12 1012 2012 3012 4012 5012 6012 7012 8012  9012
#>   [13,] 1   13 1013 2013 3013 4013 5013 6013 7013 8013  9013
#>   [14,] 1   14 1014 2014 3014 4014 5014 6014 7014 8014  9014
#>   [15,] 1   15 1015 2015 3015 4015 5015 6015 7015 8015  9015
#>   [16,] 1   16 1016 2016 3016 4016 5016 6016 7016 8016  9016
#>   [17,] 1   17 1017 2017 3017 4017 5017 6017 7017 8017  9017
#>   [18,] 1   18 1018 2018 3018 4018 5018 6018 7018 8018  9018
#>   [19,] 1   19 1019 2019 3019 4019 5019 6019 7019 8019  9019
#>   [20,] 1   20 1020 2020 3020 4020 5020 6020 7020 8020  9020
#>   [21,] 1   21 1021 2021 3021 4021 5021 6021 7021 8021  9021
#>   [22,] 1   22 1022 2022 3022 4022 5022 6022 7022 8022  9022
#>   [23,] 1   23 1023 2023 3023 4023 5023 6023 7023 8023  9023
#>   [24,] 1   24 1024 2024 3024 4024 5024 6024 7024 8024  9024
#>   [25,] 1   25 1025 2025 3025 4025 5025 6025 7025 8025  9025
#>   [26,] 1   26 1026 2026 3026 4026 5026 6026 7026 8026  9026
#>   [27,] 1   27 1027 2027 3027 4027 5027 6027 7027 8027  9027
#>   [28,] 1   28 1028 2028 3028 4028 5028 6028 7028 8028  9028
#>   [29,] 1   29 1029 2029 3029 4029 5029 6029 7029 8029  9029
#>   [30,] 1   30 1030 2030 3030 4030 5030 6030 7030 8030  9030
#>   [31,] 1   31 1031 2031 3031 4031 5031 6031 7031 8031  9031
#>   [32,] 1   32 1032 2032 3032 4032 5032 6032 7032 8032  9032
#>   [33,] 1   33 1033 2033 3033 4033 5033 6033 7033 8033  9033
#>   [34,] 1   34 1034 2034 3034 4034 5034 6034 7034 8034  9034
#>   [35,] 1   35 1035 2035 3035 4035 5035 6035 7035 8035  9035
#>   [36,] 1   36 1036 2036 3036 4036 5036 6036 7036 8036  9036
#>   [37,] 1   37 1037 2037 3037 4037 5037 6037 7037 8037  9037
#>   [38,] 1   38 1038 2038 3038 4038 5038 6038 7038 8038  9038
#>   [39,] 1   39 1039 2039 3039 4039 5039 6039 7039 8039  9039
#>   [40,] 1   40 1040 2040 3040 4040 5040 6040 7040 8040  9040
#>   [41,] 1   41 1041 2041 3041 4041 5041 6041 7041 8041  9041
#>   [42,] 1   42 1042 2042 3042 4042 5042 6042 7042 8042  9042
#>   [43,] 1   43 1043 2043 3043 4043 5043 6043 7043 8043  9043
#>   [44,] 1   44 1044 2044 3044 4044 5044 6044 7044 8044  9044
#>   [45,] 1   45 1045 2045 3045 4045 5045 6045 7045 8045  9045
#>   [46,] 1   46 1046 2046 3046 4046 5046 6046 7046 8046  9046
#>   [47,] 1   47 1047 2047 3047 4047 5047 6047 7047 8047  9047
#>   [48,] 1   48 1048 2048 3048 4048 5048 6048 7048 8048  9048
#>   [49,] 1   49 1049 2049 3049 4049 5049 6049 7049 8049  9049
#>   [50,] 1   50 1050 2050 3050 4050 5050 6050 7050 8050  9050
#>   [51,] 1   51 1051 2051 3051 4051 5051 6051 7051 8051  9051
#>   [52,] 1   52 1052 2052 3052 4052 5052 6052 7052 8052  9052
#>   [53,] 1   53 1053 2053 3053 4053 5053 6053 7053 8053  9053
#>   [54,] 1   54 1054 2054 3054 4054 5054 6054 7054 8054  9054
#>   [55,] 1   55 1055 2055 3055 4055 5055 6055 7055 8055  9055
#>   [56,] 1   56 1056 2056 3056 4056 5056 6056 7056 8056  9056
#>   [57,] 1   57 1057 2057 3057 4057 5057 6057 7057 8057  9057
#>   [58,] 1   58 1058 2058 3058 4058 5058 6058 7058 8058  9058
#>   [59,] 1   59 1059 2059 3059 4059 5059 6059 7059 8059  9059
#>   [60,] 1   60 1060 2060 3060 4060 5060 6060 7060 8060  9060
#>   [61,] 1   61 1061 2061 3061 4061 5061 6061 7061 8061  9061
#>   [62,] 1   62 1062 2062 3062 4062 5062 6062 7062 8062  9062
#>   [63,] 1   63 1063 2063 3063 4063 5063 6063 7063 8063  9063
#>   [64,] 1   64 1064 2064 3064 4064 5064 6064 7064 8064  9064
#>   [65,] 1   65 1065 2065 3065 4065 5065 6065 7065 8065  9065
#>   [66,] 1   66 1066 2066 3066 4066 5066 6066 7066 8066  9066
#>   [67,] 1   67 1067 2067 3067 4067 5067 6067 7067 8067  9067
#>   [68,] 1   68 1068 2068 3068 4068 5068 6068 7068 8068  9068
#>   [69,] 1   69 1069 2069 3069 4069 5069 6069 7069 8069  9069
#>   [70,] 1   70 1070 2070 3070 4070 5070 6070 7070 8070  9070
#>   [71,] 1   71 1071 2071 3071 4071 5071 6071 7071 8071  9071
#>   [72,] 1   72 1072 2072 3072 4072 5072 6072 7072 8072  9072
#>   [73,] 1   73 1073 2073 3073 4073 5073 6073 7073 8073  9073
#>   [74,] 1   74 1074 2074 3074 4074 5074 6074 7074 8074  9074
#>   [75,] 1   75 1075 2075 3075 4075 5075 6075 7075 8075  9075
#>   [76,] 1   76 1076 2076 3076 4076 5076 6076 7076 8076  9076
#>   [77,] 1   77 1077 2077 3077 4077 5077 6077 7077 8077  9077
#>   [78,] 1   78 1078 2078 3078 4078 5078 6078 7078 8078  9078
#>   [79,] 1   79 1079 2079 3079 4079 5079 6079 7079 8079  9079
#>   [80,] 1   80 1080 2080 3080 4080 5080 6080 7080 8080  9080
#>   [81,] 1   81 1081 2081 3081 4081 5081 6081 7081 8081  9081
#>   [82,] 1   82 1082 2082 3082 4082 5082 6082 7082 8082  9082
#>   [83,] 1   83 1083 2083 3083 4083 5083 6083 7083 8083  9083
#>   [84,] 1   84 1084 2084 3084 4084 5084 6084 7084 8084  9084
#>   [85,] 1   85 1085 2085 3085 4085 5085 6085 7085 8085  9085
#>   [86,] 1   86 1086 2086 3086 4086 5086 6086 7086 8086  9086
#>   [87,] 1   87 1087 2087 3087 4087 5087 6087 7087 8087  9087
#>   [88,] 1   88 1088 2088 3088 4088 5088 6088 7088 8088  9088
#>   [89,] 1   89 1089 2089 3089 4089 5089 6089 7089 8089  9089
#>   [90,] 1   90 1090 2090 3090 4090 5090 6090 7090 8090  9090
#>  [ reached getOption("max.print") -- omitted 910 rows ]