Simulated data sets to illustrate the package functionality

Both the tweedietraindata and tweedietestdata dataframe are synthetically generated data sets to illustrate the functionality of the package. The tweedietraindata has 250 000 observations and the tweedietestdata has 250 000 observations. The same settings were used to generate both data sets.

Usage

data(tweedietraindata)
  data(tweedietestdata)

Format

y

the tweedie distributed outcome variable

cluster

the cluster

subcluster

the subcluster nested within cluster

x1

covariate 1

x2

covariate 2

x3

covariate 3

x4

covariate 4

x5

covariate 5

Details

See the examples for how the data sets were generated.

Examples

  # The data sets were generated as follows
  lapply(c("magrittr", "dplyr", "data.table", "tweedie"), library, character.only = TRUE)
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:plyr':
#> 
#>     arrange, count, desc, failwith, id, mutate, rename, summarise,
#>     summarize
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
#> 
#> Attaching package: 'data.table'
#> The following objects are masked from 'package:dplyr':
#> 
#>     between, first, last
#> [[1]]
#>  [1] "lattice"   "actuar"    "lme4"      "magrittr"  "plyr"      "statmod"  
#>  [7] "actuaRE"   "cplm"      "splines"   "Matrix"    "coda"      "stats"    
#> [13] "graphics"  "grDevices" "utils"     "datasets"  "methods"   "base"     
#> 
#> [[2]]
#>  [1] "dplyr"     "lattice"   "actuar"    "lme4"      "magrittr"  "plyr"     
#>  [7] "statmod"   "actuaRE"   "cplm"      "splines"   "Matrix"    "coda"     
#> [13] "stats"     "graphics"  "grDevices" "utils"     "datasets"  "methods"  
#> [19] "base"     
#> 
#> [[3]]
#>  [1] "data.table" "dplyr"      "lattice"    "actuar"     "lme4"      
#>  [6] "magrittr"   "plyr"       "statmod"    "actuaRE"    "cplm"      
#> [11] "splines"    "Matrix"     "coda"       "stats"      "graphics"  
#> [16] "grDevices"  "utils"      "datasets"   "methods"    "base"      
#> 
#> [[4]]
#>  [1] "tweedie"    "data.table" "dplyr"      "lattice"    "actuar"    
#>  [6] "lme4"       "magrittr"   "plyr"       "statmod"    "actuaRE"   
#> [11] "cplm"       "splines"    "Matrix"     "coda"       "stats"     
#> [16] "graphics"   "grDevices"  "utils"      "datasets"   "methods"   
#> [21] "base"      
#> 
  set.seed(1)

  # Simulate training data
  set.seed(1)
  nClusters    = 5
  nSubclusters = 5
  p            = 5
  Uj           = scale(rnorm(nClusters))
  Ujk          = do.call("c", lapply(seq_len(nClusters), function(x) scale(rnorm(nSubclusters))))
  nPop         = 1e6
  nSample      = 50
  nTest        = 1e3
  X            = replicate(p, rnorm(nPop))
  Beta         = rnorm(p)
  cluster      = sample(seq_len(nClusters), nPop, TRUE)
  subcluster   = NULL
  uniqueCl     = sort(unique(cluster))
  for(cl in uniqueCl)
    subcluster[cluster == cl] <- sample(
      1 - seq_len(nSubclusters) + which(cl == uniqueCl) * nSubclusters,
      sum(cluster == cl),
      TRUE)
  table(subcluster, cluster)
#>           cluster
#> subcluster     1     2     3     4     5
#>         1  40325     0     0     0     0
#>         2  39762     0     0     0     0
#>         3  40303     0     0     0     0
#>         4  40112     0     0     0     0
#>         5  40179     0     0     0     0
#>         6      0 39901     0     0     0
#>         7      0 39885     0     0     0
#>         8      0 39842     0     0     0
#>         9      0 39845     0     0     0
#>         10     0 39912     0     0     0
#>         11     0     0 40121     0     0
#>         12     0     0 39891     0     0
#>         13     0     0 40108     0     0
#>         14     0     0 40373     0     0
#>         15     0     0 39613     0     0
#>         16     0     0     0 40045     0
#>         17     0     0     0 39915     0
#>         18     0     0     0 40026     0
#>         19     0     0     0 39937     0
#>         20     0     0     0 40039     0
#>         21     0     0     0     0 40222
#>         22     0     0     0     0 39845
#>         23     0     0     0     0 39643
#>         24     0     0     0     0 39934
#>         25     0     0     0     0 40222
  eta       = X %*% Beta + Uj[match(cluster, seq_len(nClusters))] +
              Ujk[match(subcluster, seq_len(nClusters * nSubclusters))]
  y         = rtweedie(nPop, mu = exp(as.vector(eta)), phi = 1, power = 1.5)
  wt        = runif(nPop)
  Dt        = data.frame(y, X, wt, cluster, subcluster)
  colnames(Dt) %<>% tolower

  tweedietraindata = Dt %>%
    group_by(subcluster) %>%
    sample_n(size = nSample) %>%
    as.data.table

  tweedietestdata = Dt %>%
    group_by(subcluster) %>%
    sample_n(size = nSample) %>%
    as.data.table