Hachemeister Data Set

Long format of the Hachemeister (1975) data set giving average claim amounts in private passenger bodily injury insurance. We have data of five U.S. states over 12 quarters between July 1970 and June 1973 and we have the corresponding number of claims. To obtain a hierarchical structure, we created an artificial variable cohort. With this, we created a hierarchical multi-level factor, with cohort as the first hierarchical level and state as the second hierarchical level, nested within cohort.

hachemeisterLong

Format

A data.frame with 60 rows and the following 5 columns:

cohort

artificially created variable;

state

the state number;

time

time variable (quarter of the observation);

ratio

the average claim amount;

weight

the corresponding number of claims.

Source

Hachemeister, C. A. (1975), Credibility for regression models with application to trend, Proceedings of the Berkeley Actuarial Research Conference on Credibility, Academic Press.