Combining the hierarchical credibility model with a GLM (Ohlsson, 2008)

Fit a random effects model using Ohlsson's methodology. In this function you explicitly specify the power parameter p. See hierCredTweedie when you also want to estimate the p.

hierCredGLM(
  formula,
  data,
  weights,
  p = 1.5,
  link.power = 0,
  muHatGLM = TRUE,
  epsilon = 1e-04,
  maxiter = 500,
  maxiterGLM = 500,
  verbose = FALSE,
  returnData = TRUE,
  balanceProperty = TRUE,
  y = TRUE,
  ...
)

Arguments

formula

object of type formula that specifies which model should be fitted. Syntax is the same as for lmer and glmer. For example, Yijkt ~ x1 + x2 + (1 | Industry / Branch).

data

an object that is coercible by as.data.table, containing the variables in the model.

weights

variable name of the exposure weight.

p

the value for the power parameter of the Tweedie distribution, which is passed to tweedie. Default is 1.5.

link.power

index of power link function, which is passed to tweedie. link.power = 0 produces a log-link. Defaults to the canonical link, which is 1 - p.

muHatGLM

indicates which estimate has to be used in the algorithm for the intercept term. Default is TRUE, which used the intercept as estimated by the GLM. If FALSE, the estimate of the hierarchical credibility model is used.

epsilon

positive convergence tolerance \(\epsilon\); the iterations converge when 7 \(||\theta[k] - \theta[k - 1]||^2[[2]]/||\theta[k - 1]||^2[[2]] < \epsilon\). Here, \(\theta[k]\) is the parameter vector at the \(k^{th}\) iteration.

maxiter

maximum number of iterations.

maxiterGLM

maximum number of iterations when fitting the GLM part. Passed to glm.

verbose

logical indicating if output should be produced during the algorithm.

returnData

logical indicating if input data has to be returned.

balanceProperty

logical indicating if the balance property should be satisfied.

y

logical indicating whether the response vector should be returned as a component of the returned value.

...

arguments passed to glm

Value

An object of type hierCredGLM with the following slots:

call

the matched call

HierarchicalResults

results of the hierarchical credibility model.

fitGLM

the results from fitting the GLM part.

iter

total number of iterations.

Converged

logical indicating whether the algorithm converged.

LevelsCov

object that summarizes the unique levels of each of the contract-specific covariates.

fitted.values

the fitted mean values, resulting from the model fit.

prior.weights

the weights (exposure) initially supplied.

y

if requested, the response vector. Default is TRUE.

References

Campo, B.D.C. and Antonio, Katrien (2023). Insurance pricing with hierarchically structured data an illustration with a workers' compensation insurance portfolio. Scandinavian Actuarial Journal, doi: 10.1080/03461238.2022.2161413

Ohlsson, E. (2008). Combining generalized linear models and credibility models in practice. Scandinavian Actuarial Journal 2008(4), 301–314.

See also

Examples

# \donttest{
data("tweedietraindata")
fit = hierCredGLM(y ~ x1 + (1 | cluster / subcluster), tweedietraindata, weights = wt,
p = 1.7)
fit
#> Call:
#> hierCredGLM(formula = y ~ x1 + (1 | cluster/subcluster), data = tweedietraindata, 
#>     weights = wt, p = 1.7)
#> 
#> 
#> Combination of the hierarchical credibility model with a GLM
#> 
#> Estimated variance parameters:
#>   Var(V[jk]): 10.01644 
#>   Var(V[j]): 15.40687 
#> Unique number of categories of cluster: 5
#> Unique number of categories of subcluster: 25
#> 
#> Results contract-specific risk factors:
#> 
#> 
#> Call:  glm(formula = FormulaGLM, family = tweedie(var.power = p, link.power = 0), 
#>     data = data, weights = data$wijkt, model = TRUE, y = TRUE)
#> 
#> Coefficients:
#> (Intercept)           x1  
#>      1.0868      -0.2541  
#> 
#> Degrees of Freedom: 1249 Total (i.e. Null);  1248 Residual
#> Null Deviance:	    1827 
#> Residual Deviance: 1782 	AIC: NA
summary(fit)
#> Call:
#> hierCredGLM(formula = y ~ x1 + (1 | cluster/subcluster), data = tweedietraindata, 
#>     weights = wt, p = 1.7)
#> 
#> 
#> Combination of the hierarchical credibility model with a GLM
#> 
#> Estimated variance parameters:
#>   Individual contracts: 28.97471 
#>   Var(V[jk]): 10.01644 
#>   Var(V[j]): 15.40687 
#> Unique number of categories of cluster: 5
#> Unique number of categories of subcluster: 25
#> 
#> Results contract-specific risk factors:
#> 
#> 
#> Call:
#> glm(formula = FormulaGLM, family = tweedie(var.power = p, link.power = 0), 
#>     data = data, weights = data$wijkt, model = TRUE, y = TRUE)
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  1.08680    0.04496  24.174  < 2e-16 ***
#> x1          -0.25406    0.04614  -5.507 4.44e-08 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> (Dispersion parameter for Tweedie family taken to be 1.549018)
#> 
#>     Null deviance: 1827  on 1249  degrees of freedom
#> Residual deviance: 1782  on 1248  degrees of freedom
#> AIC: NA
#> 
#> Number of Fisher Scoring iterations: 5
#> 
ranef(fit)
#> $sector
#> Key: <cluster>
#>    cluster        Uj
#>      <int>     <num>
#> 1:       1 0.4639580
#> 2:       2 1.1053967
#> 3:       3 0.4051336
#> 4:       4 3.4083454
#> 5:       5 0.8859055
#> 
#> $group
#> Key: <cluster, subcluster>
#>     cluster subcluster       Ujk
#>       <int>      <num>     <num>
#>  1:       1          1 0.1861397
#>  2:       1          2 0.7960008
#>  3:       1          3 1.8820107
#>  4:       1          4 0.9922363
#>  5:       1          5 0.3924754
#>  6:       2          6 2.6063843
#>  7:       2          7 0.9708932
#>  8:       2          8 0.4218217
#>  9:       2          9 0.2826499
#> 10:       2         10 0.7802388
#> 11:       3         11 0.1901278
#> 12:       3         12 0.3081920
#> 13:       3         13 1.4666131
#> 14:       3         14 1.1745657
#> 15:       3         15 0.9059039
#> 16:       4         16 1.2140636
#> 17:       4         17 1.7185047
#> 18:       4         18 0.7812095
#> 19:       4         19 0.2908161
#> 20:       4         20 1.4547884
#> 21:       5         21 1.3261554
#> 22:       5         22 0.7187670
#> 23:       5         23 0.1867345
#> 24:       5         24 0.7030996
#> 25:       5         25 1.9815144
#>     cluster subcluster       Ujk
#>       <int>      <num>     <num>
#> 
fixef(fit)
#> (Intercept)          x1 
#>    1.086799   -0.254055 
# }