Combining the hierarchical credibility model with a GLM (Ohlsson, 2008)

Fit a random effects model using Ohlsson's methodology. In this function you estimate the power parameter p. See hierCredGLM when you want fix p.

hierCredTweedie(
  formula,
  data,
  weights,
  muHatGLM = TRUE,
  epsilon = 1e-04,
  maxiter = 500,
  verbose = FALSE,
  returnData = TRUE,
  cpglmControl = list(bound.p = c(1.01, 1.99)),
  balanceProperty = TRUE,
  optimizer = "bobyqa",
  y = TRUE,
  ...
)

Arguments

formula

object of type formula that specifies which model should be fitted. Syntax is the same as for lmer and glmer. For example, Yijkt ~ x1 + x2 + (1 | Industry / Branch).

data

an object that is coercible by as.data.table, containing the variables in the model.

weights

variable name of the exposure weight.

muHatGLM

indicates which estimate has to be used in the algorithm for the intercept term. Default is TRUE, which used the intercept as estimated by the GLM. If FALSE, the estimate of the hierarchical credibility model is used.

epsilon

positive convergence tolerance \(\epsilon\); the iterations converge when \(||\theta[k] - \theta[k - 1]||^2[[2]]/||\theta[k - 1]||^2[[2]] < \epsilon\). Here, \(\theta[k]\) is the parameter vector at the \(k^{th}\) iteration.

maxiter

maximum number of iterations.

verbose

logical indicating if output should be produced during the algorithm.

returnData

logical indicating if input data has to be returned.

cpglmControl

a list of parameters to control the fitting process in the GLM part. By default, cpglmControl = list(bound.p = c(1.01, 1.99)) which restricts the range of the power parameter p to [1.01, 1.99] in the fitting process. This list is passed to cpglm.

balanceProperty

logical indicating if the balance property should be satisfied.

optimizer

a character string that determines which optimization routine is to be used in estimating the index and the dispersion parameters. Possible choices are "nlminb" (the default, see nlminb), "bobyqa" (bobyqa) and "L-BFGS-B" (optim).

y

logical indicating whether the response vector should be returned as a component of the returned value.

...

arguments passed to cpglm.

Value

An object of type hierCredTweedie with the following slots:

call

the matched call

HierarchicalResults

results of the hierarchical credibility model.

fitGLM

the results from fitting the GLM part.

iter

total number of iterations.

Converged

logical indicating whether the algorithm converged.

LevelsCov

object that summarizes the unique levels of each of the contract-specific covariates.

fitted.values

the fitted mean values, resulting from the model fit.

prior.weights

the weights (exposure) initially supplied.

y

if requested, the response vector. Default is TRUE.

Details

When estimating the GLM part, this function uses the cpglm function from the cplm package.

References

Campo, B.D.C. and Antonio, Katrien (2023). Insurance pricing with hierarchically structured data an illustration with a workers' compensation insurance portfolio. Scandinavian Actuarial Journal, doi: 10.1080/03461238.2022.2161413

Ohlsson, E. (2008). Combining generalized linear models and credibility models in practice. Scandinavian Actuarial Journal 2008(4), 301–314.

See also

Examples

# \donttest{
data("tweedietraindata")
fit = hierCredTweedie(y ~ x1 + (1 | cluster / subcluster), tweedietraindata, weights = wt,
p = 1.7)
fit
#> Call:
#> hierCredTweedie(formula = y ~ x1 + (1 | cluster/subcluster), 
#>     data = tweedietraindata, weights = wt, p = 1.7)
#> 
#> 
#> Combination of the hierarchical credibility model with a GLM
#> 
#> Estimated variance parameters:
#>   Var(V[jk]): 10.01849 
#>   Var(V[j]): 15.40409 
#> Unique number of categories of cluster: 5
#> Unique number of categories of subcluster: 25
#> 
#> Results contract-specific risk factors:
#> 
#> 
#> Call:
#> cpglm(formula = FormulaGLM, link = "log", data = data, weights = wijkt, 
#>     control = cpglmControl, optimizer = optimizer, p = 1.7)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -3.3729  -1.2209  -0.4933   0.0401   4.3885  
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  1.14586    0.04508   25.42  < 2e-16 ***
#> x1          -0.25468    0.04630   -5.50 4.59e-08 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Estimated dispersion parameter: 1.0912
#> Estimated index parameter: 1.7133 
#> 
#> Residual deviance: 1808.1  on 1248  degrees of freedom
#> AIC:  5823.8 
#> 
#> Number of Fisher Scoring iterations:  5 
#> 
summary(fit)
#> Call:
#> hierCredTweedie(formula = y ~ x1 + (1 | cluster/subcluster), 
#>     data = tweedietraindata, weights = wt, p = 1.7)
#> 
#> 
#> Combination of the hierarchical credibility model with a GLM
#> 
#> Estimated variance parameters:
#>   Individual contracts: 28.96287 
#>   Var(V[jk]): 10.01849 
#>   Var(V[j]): 15.40409 
#> Unique number of categories of cluster: 5
#> Unique number of categories of subcluster: 25
#> 
#> Results contract-specific risk factors:
#> 
#> 
#> Call:
#> cpglm(formula = FormulaGLM, link = "log", data = data, weights = wijkt, 
#>     control = cpglmControl, optimizer = optimizer, p = 1.7)
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -3.3729  -1.2209  -0.4933   0.0401   4.3885  
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  1.14586    0.04508   25.42  < 2e-16 ***
#> x1          -0.25468    0.04630   -5.50 4.59e-08 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Estimated dispersion parameter: 1.0912
#> Estimated index parameter: 1.7133 
#> 
#> Residual deviance: 1808.1  on 1248  degrees of freedom
#> AIC:  5823.8 
#> 
#> Number of Fisher Scoring iterations:  5 
#> 
#> $call
#> cpglm(formula = FormulaGLM, link = "log", data = data, weights = wijkt, 
#>     control = cpglmControl, optimizer = optimizer, p = 1.7)
#> 
#> $deviance
#> [1] 1808.058
#> 
#> $aic
#> [1] 5823.829
#> 
#> $contrasts
#> NULL
#> 
#> $df.residual
#> [1] 1248
#> 
#> $iter
#> [1] 5
#> 
#> $na.action
#> NULL
#> 
#> $deviance.resid
#>         Min          1Q      Median          3Q         Max 
#> -3.37290029 -1.22089715 -0.49332591  0.04012825  4.38852057 
#> 
#> $coefficients
#>               Estimate Std. Error   t value      Pr(>|t|)
#> (Intercept)  1.1458637 0.04507941 25.418782 3.334576e-115
#> x1          -0.2546834 0.04630250 -5.500424  4.592009e-08
#> 
#> $dispersion
#> [1] 1.091179
#> 
#> $vcov
#>             (Intercept)          x1
#> (Intercept) 0.002032153 0.000124068
#> x1          0.000124068 0.002143922
#> 
#> $p
#> [1] 1.713276
#> 
ranef(fit)
#> $sector
#> Key: <cluster>
#>    cluster        Uj
#>      <int>     <num>
#> 1:       1 0.4443078
#> 2:       2 1.0491740
#> 3:       3 0.3890435
#> 4:       4 3.2173266
#> 5:       5 0.8414421
#> 
#> $group
#> Key: <cluster, subcluster>
#>     cluster subcluster       Ujk
#>       <int>      <num>     <num>
#>  1:       1          1 0.1848060
#>  2:       1          2 0.7834655
#>  3:       1          3 1.8538499
#>  4:       1          4 0.9766869
#>  5:       1          5 0.3877684
#>  6:       2          6 2.5910069
#>  7:       2          7 0.9638886
#>  8:       2          8 0.4194167
#>  9:       2          9 0.2814559
#> 10:       2         10 0.7747147
#> 11:       3         11 0.1882565
#> 12:       3         12 0.3037132
#> 13:       3         13 1.4433668
#> 14:       3         14 1.1531475
#> 15:       3         15 0.8901571
#> 16:       4         16 1.2112862
#> 17:       4         17 1.7150803
#> 18:       4         18 0.7796862
#> 19:       4         19 0.2905974
#> 20:       4         20 1.4515797
#> 21:       5         21 1.3155349
#> 22:       5         22 0.7138866
#> 23:       5         23 0.1861284
#> 24:       5         24 0.6981412
#> 25:       5         25 1.9637542
#>     cluster subcluster       Ujk
#>       <int>      <num>     <num>
#> 
fixef(fit)
#> (Intercept)          x1 
#>   1.1458637  -0.2546834 
# }