Combining the hierarchical credibility model with a GLM (Ohlsson, 2008)

Fit a random effects model using Ohlsson's methodology. In this function you estimate the power parameter p. See hierCredGLM when you want fix p.

Usage

hierCredTweedie(
  formula,
  data,
  weights,
  muHatGLM = T,
  epsilon = 1e-04,
  maxiter = 500,
  verbose = F,
  returnData = T,
  cpglmControl = list(bound.p = c(1.01, 1.99)),
  balanceProperty = T,
  optimizer = "bobyqa",
  y = T,
  ...
)

Arguments

formula

object of type formula that specifies which model should be fitted. Syntax is the same as for lmer and glmer. For example, Yijkt ~ x1 + x2 + (1 | Industry / Branch).

data

an object that is coercible by as.data.table, containing the variables in the model.

weights

variable name of the exposure weight.

muHatGLM

indicates which estimate has to be used in the algorithm for the intercept term. Default is TRUE, which used the intercept as estimated by the GLM. If FALSE, the estimate of the hierarchical credibility model is used.

epsilon

positive convergence tolerance \(\epsilon\); the iterations converge when \(||\theta[k] - \theta[k - 1]||^2[[2]]/||\theta[k - 1]||^2[[2]] < \epsilon\). Here, \(\theta[k]\) is the parameter vector at the \(k^{th}\) iteration.

maxiter

maximum number of iterations.

verbose

logical indicating if output should be produced during the algorithm.

returnData

logical indicating if input data has to be returned.

cpglmControl

a list of parameters to control the fitting process in the GLM part. By default, cpglmControl = list(bound.p = c(1.01, 1.99)) which restricts the range of the power parameter p to [1.01, 1.99] in the fitting process. This list is passed to cpglm.

balanceProperty

logical indicating if the balance property should be satisfied.

optimizer

a character string that determines which optimization routine is to be used in estimating the index and the dispersion parameters. Possible choices are "nlminb" (the default, see nlminb), "bobyqa" (bobyqa) and "L-BFGS-B" (optim).

y

logical indicating whether the response vector should be returned as a component of the returned value.

...

arguments passed to cpglm.

Value

An object of type hierCredTweedie with the following slots:

call

the matched call

HierarchicalResults

results of the hierarchical credibility model.

fitGLM

the results from fitting the GLM part.

iter

total number of iterations.

Converged

logical indicating whether the algorithm converged.

LevelsCov

object that summarizes the unique levels of each of the contract-specific covariates.

fitted.values

the fitted mean values, resulting from the model fit.

prior.weights

the weights (exposure) initially supplied.

y

if requested, the response vector. Default is TRUE.

Details

When estimating the GLM part, this function uses the cpglm function from the cplm package.

References

Ohlsson, E. (2008). Combining generalized linear models and credibility models in practice. Scandinavian Actuarial Journal 2008(4), 301–314.

See also

Examples

if (FALSE) {
data("dataCar")
fit = hierCredTweedie(Y ~ area + (1 | VehicleType / VehicleBody), dataCar, weights = w, epsilon = 1e-6)
fit
summary(fit)
ranef(fit)
fixef(fit)
}